+++ tests/der_constructed.rs
@@ -31,6 +31,27 @@ fn tc_der_tagged_explicit_g(i: &[u8], out: Result<u32, BerError>) {
     }
 }
 
+#[test_case(&hex!("82 03 01 00 01"), Ok(0x10001) ; "tag ok")]
+#[test_case(&hex!("83 03 01 00 01"), Err(BerError::InvalidTag) ; "invalid tag")]
+fn tc_der_tagged_implicit_g(i: &[u8], out: Result<u32, BerError>) {
+    fn parse_int_implicit(i: &[u8]) -> BerResult<u32> {
+        parse_der_tagged_implicit_g(2, |content, hdr, depth| {
+            let (rem, obj) = parse_der_content(DerTag::Integer)(content, &hdr, depth)?;
+            let value = obj.as_u32()?;
+            Ok((rem, value))
+        })(i)
+    }
+    let res = parse_int_implicit(i);
+    match out {
+        Ok(expected) => {
+            pretty_assertions::assert_eq!(res, Ok((&b""[..], expected)));
+        }
+        Err(e) => {
+            pretty_assertions::assert_eq!(res, Err(Err::Error(e)));
+        }
+    }
+}
+
 #[test_case(&hex!("30 00"), Ok(&[]) ; "empty seq")]
 #[test_case(&hex!("30 0a 02 03 01 00 01 02 03 01 00 00"), Ok(&[0x10001, 0x10000]) ; "seq ok")]
 #[test_case(&hex!("30 07 02 03 01 00 01 02 03 01"), Err(BerError::NomError(ErrorKind::Eof)) ; "incomplete")]